apps/worker/requirements.txt
# Task Queue
celery[redis]==5.3.4
redis==4.6.0

# Document Processing
PyPDF2==3.0.1
pdfplumber==0.10.3
python-docx==1.1.0

# AI/ML
openai==1.6.1
openai-whisper==20231117
torch==2.1.2
transformers==4.36.2
sentence-transformers==2.2.2

# Vector Store
qdrant-client==1.7.0

# Graph Database
neo4j==5.15.0

# Utilities
python-dotenv==1.0.0
httpx==0.25.2
pillow==10.1.0
